Minor refactor of tests
