+++ core/src/test/java/io/confluent/rest/ApplicationServerTest.java
@@ -2,6 +2,7 @@ package io.confluent.rest;
 
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.net.URL;
 
 import org.apache.kafka.common.config.ConfigException;
 import org.apache.http.client.methods.CloseableHttpResponse;
@@ -12,6 +13,7 @@ import org.eclipse.jetty.http.HttpStatus;
 import org.eclipse.jetty.http.HttpStatus.Code;
 import org.eclipse.jetty.util.resource.Resource;
 import org.eclipse.jetty.util.resource.ResourceCollection;
+import org.eclipse.jetty.server.ServerConnector;
 import org.glassfish.jersey.servlet.ServletProperties;
 import org.junit.After;
 import org.junit.Before;
@@ -20,10 +22,12 @@ import org.junit.Test;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.stream.Collectors;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
@@ -127,9 +131,30 @@ public class ApplicationServerTest {
     assertThat(makeGetRequest("/app2/index.html"), is(Code.OK));
   }
 
+  List<URL> getListeners() {
+    return Arrays.stream(server.getConnectors())
+            .filter(connector -> connector instanceof ServerConnector)
+            .map(ServerConnector.class::cast)
+            .map(connector -> {
+              try {
+                final String protocol = new HashSet<>(connector.getProtocols())
+                        .stream()
+                        .map(String::toLowerCase)
+                        .anyMatch(s -> s.equals("ssl")) ? "https" : "http";
+
+                final int localPort = connector.getLocalPort();
+
+                return new URL(protocol, "localhost", localPort, "");
+              } catch (final Exception e) {
+                throw new RuntimeException("Malformed listener", e);
+              }
+            })
+            .collect(Collectors.toList());
+  }
+
   @SuppressWarnings("SameParameterValue")
   private HttpStatus.Code makeGetRequest(final String path) throws Exception {
-    final HttpGet httpget = new HttpGet(server.getListeners().get(0).toString() + path);
+    final HttpGet httpget = new HttpGet(getListeners().get(0).toString() + path);
 
     try (CloseableHttpClient httpClient = HttpClients.createDefault();
          CloseableHttpResponse response = httpClient.execute(httpget)) {
